Integrate okhttp3 instrumentation with a separate plugin
