+++ agent/build.gradle.kts
@@ -24,7 +24,8 @@ dependencies {
     // TODO api(project(":integration:session-recording"))
     api(project(":integration:crash"))
     api(project(":integration:anr"))
-    api(project(":integration:networkrequest"))
+    api(project(":integration:httpurlconnection"))
+    api(project(":integration:okhttp3"))
     api(project(":integration:startup"))
     api(project(":integration:interactions"))
     api(project(":instrumentation:runtime:customtracking"))
