+++ integration/httpurlconnection/src/main/assets/LICENSE.txt
