+++ integration/okhttp3/build.gradle.kts
@@ -1,0 +1,34 @@
+import plugins.ConfigAndroidLibrary
+import plugins.ConfigPublish
+import utils.artifactIdProperty
+import utils.artifactPrefix
+import utils.integrationPrefix
+import utils.versionProperty
+
+plugins {
+    id("com.android.library")
+    id("kotlin-android")
+}
+
+apply<ConfigAndroidLibrary>()
+apply<ConfigPublish>()
+
+ext {
+    set(artifactIdProperty, "$artifactPrefix$integrationPrefix${project.name}")
+    set(versionProperty, Configurations.sdkVersionName)
+}
+
+android {
+    namespace = "com.splunk.rum.integration.okhttp3"
+}
+
+dependencies {
+    api(platform(Dependencies.Otel.androidBom))
+
+    implementation(project(":integration:agent:internal"))
+
+    implementation(Dependencies.Otel.androidOkHttp3Library)
+
+    implementation(Dependencies.SessionReplay.commonLogger)
+    implementation(Dependencies.SessionReplay.commonUtils)
+}
